First pass at pulling out InstrumentSettings form fields into separate component
